Implemented an Editor Window for generating random AES key and IV.
Switched to generating random AES Key and IV instead of using a password.
Noticed a significant number of 'A' characters at the end of the string, typically around 9 to 10, in the generated key, likely due to zero padding, making the password weaker.
Therefore, I proceeded with random password generation for  stronger security.